HSEARCH-2219 Move ScopedLuceneAnalyzer to org.hibernate.search.analyzer.impl
It makes sense now that ScopedAnalyzer is located in org.hibernate.search.analyzer.spi.
