+++ engine/src/main/java/org/hibernate/search/backend/spi/SingularTermDeletionQuery.java
@@ -18,12 +18,12 @@ import org.apache.lucene.search.TermQuery;
 import org.hibernate.search.analyzer.spi.AnalyzerReference;
 import org.hibernate.search.analyzer.impl.LuceneAnalyzerReference;
 import org.hibernate.search.analyzer.impl.RemoteAnalyzerReference;
+import org.hibernate.search.analyzer.impl.ScopedLuceneAnalyzer;
 import org.hibernate.search.bridge.FieldBridge;
 import org.hibernate.search.bridge.util.impl.ContextualExceptionBridgeHelper;
 import org.hibernate.search.bridge.util.impl.NumericFieldUtils;
 import org.hibernate.search.engine.spi.DocumentBuilderIndexedEntity;
 import org.hibernate.search.exception.AssertionFailure;
-import org.hibernate.search.util.impl.ScopedLuceneAnalyzer;
 
 /**
  * DeleteByQuery equivalent to {@link org.apache.lucene.search.TermQuery}
